load fonts (untested)

+++ src/grp.rs
@@ -22,6 +22,18 @@ impl GRPHeader {
     }
 }
 
+// all the frames in the GRP file must be the exact same size a lot of the
+// time, frames come in matching sets of 17. In the first frame, the unit
+// faces north and in the seventeenth frame it faces south, and all the frames
+// in between are of the unit rotating clockwise through the 180 degrees
+// between north and south. Since there are only frames of the unit pointing
+// north, south and east, you probably have already worked out that StarCraft
+// flips the graphics to display graphics pointing west. Not all units are
+// like this though. Only if the images.dat entry for the GRP file has the
+// "GFX Turns" property set to 1 in Arsenal Zero will StarCraft expect sets of
+// 17 rotated frames.
+//
+// no color palette is saved in the GRP file
 pub struct GRP {
     pub header: GRPHeader,
     pub frames: Vec<Vec<u8>>,
@@ -70,14 +82,17 @@ impl GRP {
         // read the actual frame data
         // i.e. offsets to rle encoded line data beginnings
         for frame_int in &frames_int {
-            let line_offsets = GRP::read_line_offsets(file, frame_int.frameoffset,
-                                                      frame_int.frameheight as usize);
+            let line_offsets =
+                GRP::read_line_offsets(file, frame_int.frameoffset, frame_int.frameheight as usize);
             let mut frame_data = vec![0 as u8; (header.width as usize * header.width as usize)];
             for i in 0..line_offsets.len() {
-                GRP::read_line_data(file, frame_int.y_offset as usize + i,
+                GRP::read_line_data(file,
+                                    frame_int.y_offset as usize + i,
                                     frame_int.frameoffset as u64 + line_offsets[i] as u64,
-                                    frame_int.x_offset as usize, frame_int.framewidth,
-                                    &mut frame_data, header.width);
+                                    frame_int.x_offset as usize,
+                                    frame_int.framewidth,
+                                    &mut frame_data,
+                                    header.width);
             }
 
             frames.push(frame_data);
@@ -86,8 +101,12 @@ impl GRP {
         frames
     }
 
-    pub fn read_line_data<T: Read + Seek>(file: &mut T, line_idx: usize, line_offset: u64,
-                                          xoffset: usize, framewidth: u8, data: &mut Vec<u8>,
+    pub fn read_line_data<T: Read + Seek>(file: &mut T,
+                                          line_idx: usize,
+                                          line_offset: u64,
+                                          xoffset: usize,
+                                          framewidth: u8,
+                                          data: &mut Vec<u8>,
                                           real_framewidth: u16) {
         file.seek(SeekFrom::Start(line_offset)).ok();
         let data_start = line_idx * real_framewidth as usize;
@@ -101,7 +120,7 @@ impl GRP {
             } else if val >= 64 {
                 // repeat the next byte val - 64 times
                 let next_val = file.read_u8().unwrap();
-                for _ in 0..val-64 {
+                for _ in 0..val - 64 {
                     data[data_start + x] = next_val;
                     x = x + 1;
                 }
@@ -131,13 +150,10 @@ impl GRP {
         outfile.write_fmt(format_args!("{0} {1}\n", self.header.width, self.header.height)).ok();
         outfile.write_fmt(format_args!("255\n")).ok();
         let frame = &self.frames[frame];
-        for i in 0..(self.header.width as usize)*(self.header.height as usize) {
-            let pal_idx: usize = 3 * (frame[i as usize] as usize);
-            outfile.write_fmt(format_args!("{0} {1} {2}\n",
-                                           pal_idx,
-                                           pal_idx,
-                                           pal_idx)).ok();
+        for i in 0..(self.header.width as usize) * (self.header.height as usize) {
+            let pal_idx: usize = 3 * (frame[i] as usize);
+            outfile.write_fmt(format_args!("{0} {1} {2}\n", pal_idx, pal_idx, pal_idx))
+                .ok();
         }
     }
 }
-
